Restrict implicit export of final mapping expression (fix #91)
BREAKING CHANGE: In .js/.coffee mapping files, for the final expression
to be implicitly exported, it must be an object expression (literal) or
function expression.

The idea is that such an expression wouldn't do much by itself.
(In particular, it cannot be an assignment or function call.
We were getting too many false positives in this case.)