Improve when onInit callbacks get called
Previously, every ')' or end of directory recursion triggered a re-init
of every mapping, which isn't necessary if e.g. there aren't any more
files to process, or we're about to dip back into a directory.

Now we delay checks for changes in the init list until just before we
actually need to: when loading or initializing a new input.
We make the two init lists identical objects when they match,
so that these tests are usually cheap.